# lib/api-zod/src/generated/types/adminAuditLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAuditLogEntryPayload } from './adminAuditLogEntryPayload';

export interface AdminAuditLogEntry {
  id: string;
  at: Date;
  /** @nullable */
  actorId?: string | null;
  /** @nullable */
  actorEmail?: string | null;
  action: string;
  /** @nullable */
  targetType?: string | null;
  /** @nullable */
  targetId?: string | null;
  payload: AdminAuditLogEntryPayload;
}