# lib/api-zod/src/generated/types/invitation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationStatus } from './invitationStatus';

export interface Invitation {
  id: string;
  email: string;
  /** @nullable */
  invitedBy?: string | null;
  suggestedRoles: string[];
  status: InvitationStatus;
  /** @nullable */
  expiresAt?: Date | null;
  createdAt: Date;
  /** @nullable */
  acceptedAt?: Date | null;
  /** @nullable */
  acceptedUserId?: string | null;
  /** @nullable */
  revokedAt?: Date | null;
  /** @nullable */
  lastSentAt?: Date | null;
}