# lib/api-zod/src/generated/types/invitationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationInput {
  /** @minLength 3 */
  email: string;
  suggestedRoles?: string[];
}