# lib/api-zod/src/generated/types/roleUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoleUpdateInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  label?: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  description?: string;
  sortOrder?: number;
}