# lib/db/src/schema/invitations.ts
import {
  pgTable,
  text,
  timestamp,
  uuid,
} from "drizzle-orm/pg-core";

export const invitationsTable = pgTable("invitations", {
  id: uuid("id").primaryKey().defaultRandom(),
  email: text("email").notNull().unique(),
  invitedBy: text("invited_by"),
  suggestedRoles: text("suggested_roles").array().notNull().default([]),
  status: text("status").notNull().default("pending"),
  expiresAt: timestamp("expires_at", { withTimezone: true }),
  createdAt: timestamp("created_at", { withTimezone: true })
    .notNull()
    .defaultNow(),
  acceptedAt: timestamp("accepted_at", { withTimezone: true }),
  acceptedUserId: text("accepted_user_id"),
  revokedAt: timestamp("revoked_at", { withTimezone: true }),
  lastSentAt: timestamp("last_sent_at", { withTimezone: true }),
});

export type Invitation = typeof invitationsTable.$inferSelect;